Fix deadlock on application close (mono)
It appears that on mono, there is a race condition when the form receives the form close message. The form tries to wait for the game agent to acknowlege a kill message, however the agent never responds (because the thread is already dead or locked up, not sure which). The simple fix is to make the kill command asynchronous so the form can close directly without waiting for the agent.